% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Basel2006_inj}
\alias{Basel2006_inj}
\title{2006 Basel EGS injection profile.}
\format{A data frame with 9310 rows and 3 variables:
\describe{
\item{t}{time (in decimal days) since 2006-12-02 00:00:00.00}
\item{dV}{volume (in cubic metres) injected in the minute preceding t}
\item{V}{cumulative volume (in cubic metres) injected up to time t}
}}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S0375650508000382}
}
\usage{
Basel2006_inj
}
\description{
A dataset containing the injected fluid volume during the 2006 stimulation
at the Basel, Switzerland, Enhanced Geothermal System. Digitzed (with 1-minute
resolution) from Fig. 5a of Haering M.O., Schanz U., Ladner F., Dyer B.C.
(2008), Characterisation of the Basel 1 enhanced geothermal system. Geothermics,
37, 469-495, doi: 10.1016/j.geothermics.2008.06.002 - Post-injection bleed-off
not included.
}
\keyword{datasets}
