% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Basel2006_seism}
\alias{Basel2006_seism}
\title{2006 Basel EGS earthquake catalogue.}
\format{A data frame with 1855 rows and 2 variables:
\describe{
\item{t}{earthquake occurrence time (in decimal days) since injection start at 2006-12-02 18:00:00.00}
\item{m}{earthquake magnitude}
}}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S0375650514000686}
}
\usage{
Basel2006_seism
}
\description{
A dataset containing the earthquakes induced during the 2006 stimulation
at the Basel, Switzerland, Enhanced Geothermal System. Catalogue filtered from
the file mmc1.txt, Appendix A. Supplementary data of Kraft T., Deichmann N. (2014),
High-precision relocation and focal mechanism of the injection-induced seismicity at
the Basel EGS. Geothermics, 52, 59-73, doi: 10.1016/j.geothermics.2014.05.014.
Time reformatted to decimal days since injection start time with only first 12 days
kept.
}
\keyword{datasets}
