% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Basel2006_seism_simulated}
\alias{Basel2006_seism_simulated}
\title{Simulated version of the 2006 Basel EGS earthquake catalogue}
\format{
A data frame with 809 rows and 2 variables:
\describe{
\item{t}{earthquake occurrence time (in decimal days) since injection start at 2006-12-02 18:00:00.00}
\item{m}{earthquake magnitude}
}
}
\usage{
Basel2006_seism_simulated
}
\description{
A dataset containing a stochastic version of the earthquakes induced during the 2006 stimulation
at the Basel, Switzerland, Enhanced Geothermal System. Simulated using the thinning method (Lewis & Shedler, 1979) with the
model parameters \code{a_fb}, \code{b}, and \code{tau} estimated from the earthquake catalogue of Kraft & Deichmann (2014)
and the injection profile \code{Basel2006_inj}. Includes events above the completeness magnitude and for 12 days
from the start of the stimulation.
}
\references{
Kraft T., Deichmann N. (2014), High-precision relocation and focal mechanism
of the injection-induced seismicity at the Basel EGS. Geothermics, 52, 59-73,
\href{https://www.sciencedirect.com/science/article/pii/S0375650514000686}{doi: 10.1016/j.geothermics.2014.05.014}

Lewis P.A.W., Shedler G.S. (1979), Simulation of nonhomogeneous Poisson processes by thinning.
Naval Res. Logistics Q., 26, 403–413, \href{https://apps.dtic.mil/dtic/tr/fulltext/u2/a059904.pdf}{doi:10.1002/nav.3800260304}
}
\seealso{
\code{Basel2006_inj}
}
\keyword{datasets}
