% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{data.bin}
\alias{data.bin}
\title{Data Binning}
\usage{
data.bin(seism, inj, dt)
}
\arguments{
\item{seism}{an earthquake catalogue data frame of parameters:
\itemize{
\item \code{t} the occurrence time (in decimal days)
\item \code{m} the magnitude
}}

\item{inj}{matching injection profile data frame of parameters:
\itemize{
\item \code{t} the occurrence time (in decimal days)
\item \code{dV} the injected volume (in cubic metres)
\item \code{V} the cumulative injected volume (in cubic metres)
}}

\item{dt}{time increment in same unit as \code{t}}
}
\value{
A list of 2 data frames binned in time:
\itemize{
\item \code{seism.rate} with parameters \code{t} and \code{rate} (number of events per \code{dt})
\item \code{inj.binned} with parameters \code{t}, \code{dV} and \code{V}
}
}
\description{
Bucketize the two main raw data sets (injection \code{inj} and seismicity \code{seism}) in \code{dt} bins
and compute the seismicity rate \code{seism.rate} and injected volume \code{inj.binned} at times t
}
\details{
Time \code{t} represents the past \code{dt} increment and is defined in the interval for which
data are available.
}
