% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{loglik_point.array}
\alias{loglik_point.array}
\title{Log Likelihood Distribution (Point Data)}
\usage{
loglik_point.array(data, par.space)
}
\arguments{
\item{data}{a list containing all the necessary point data:
\itemize{
\item \code{seism} an earthquake catalogue data frame of parameters \code{t} (time in days) and \code{m} (magnitude)
\item \code{inj} matching injection profile data frame of parameters \code{t} (time in days), \code{dV} (flow rate in cubic
metre/day) and \code{V} (cumulative injected volume in cubic metres)
\item \code{m0} the minimum magnitude cutoff of the earthquake catalogue
\item \code{ts} shut-in time (in days) (not required for "\code{injection}")
\item \code{Tmax} upper range of the time window (in days) (not required for "\code{injection}")
}}

\item{par.space}{an array of the input parameters}
}
\value{
the log likelihood distribution for the specified \code{par.space} array
}
\description{
Estimates the log likelihood distribution of the statistical model of Mignan et al. (2017) for
a specific set of input parameters. Function similar to \code{negloglik_point.val()} but for 3 vectors of input
parameters instead of 3 parameter values.
}
\details{
See Eq. A2 of Broccardo et al. (2017) for details. Used for the hierarchical Bayesian modelling.
}
\references{
Broccardo M., Mignan A., Wiemer S., Stojadinovic B., Giardini D. (2017), Hierarchical Bayesian
Modeling of Fluid‐Induced Seismicity. Geophysical Research Letters, 44 (22), 11,357-11,367,
\href{https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/2017GL075251}{doi: 10.1002/2017GL075251}

Mignan A., Broccardo M., Wiemer S., Giardini D. (2017), Induced seismicity closed-form
traffic light system for actuarial decision-making during deep fluid injections. Sci. Rep., 7, 13607,
\href{https://www.nature.com/articles/s41598-017-13585-9}{doi: 10.1038/s41598-017-13585-9}
}
\seealso{
\code{negloglik_point.val}, \code{model_posterior.distr}
}
