% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{model_par.mle_hist}
\alias{model_par.mle_hist}
\title{Model Parameter Maximum Likelihood Estimation (Histogram Data)}
\usage{
model_par.mle_hist(
  data,
  theta.init = list(a_fb = -1, tau = 1),
  window = "full sequence"
)
}
\arguments{
\item{data}{a list containing all the necessary histogram data:
\itemize{
\item \code{seism.binned} an earthquake-rate histogram data frame of parameters \code{t} (time in days) and \code{rate} (per time bin)
\item \code{inj.binned} matching binned injection profile data frame of parameters \code{t} (time in days), \code{dV} (flow rate in cubic
metre/time bin) and \code{V} (cumulative injected volume in cubic metres)
\item \code{b} the b-value of the earthquake catalogue (not required for "\code{post-injection}")
\item \code{m0} the minimum magnitude cutoff of the earthquake catalogue (not required for "\code{post-injection}")
\item \code{ts} shut-in time (in days) (not required for "\code{injection}")
\item \code{lambda0} seismicity rate (per day) at shut-in (only required for "\code{post-injection}")
}}

\item{theta.init}{an optional list of the initial values for the parameters to be optimized over
\itemize{
\item \code{a_fb} the underground feedback activation (in /cubic metre)
\item \code{tau} the mean relaxation time (in days)
}}

\item{window}{the window to be used: "\code{injection}", "\code{post-injection}", or "\verb{full sequence}"}
}
\value{
a list of the parameters' MLEs:
\itemize{
\item \code{a_fb} the underground feedback activation (in /cubic metre)
\item \code{tau} the mean relaxation time (in days)
\item \code{nLL} the negative log likelihood value for the optimized parameters
}
}
\description{
Provides the maximum likelihood estimates (MLE) of the parameters of the Non-Homogeneous Poisson Process (NHPP) derived from
the statistical model of Mignan et al. (2017) using the negative log likelihood function \code{negloglik_hist.val()} for
one of the three possible time windows (full sequence, injection phase or post-injection phase).
}
\details{
Optimization done using the `optim()`` function of the stats package. In contrast to point-data MLE parameter
optimization, the b-value is here an input instead of an output.
}
\references{
Mignan A., Broccardo M., Wiemer S., Giardini D. (2017), Induced seismicity closed-form
traffic light system for actuarial decision-making during deep fluid injections. Sci. Rep., 7, 13607,
\href{https://www.nature.com/articles/s41598-017-13585-9}{doi: 10.1038/s41598-017-13585-9}
}
\seealso{
\code{negloglik_hist.val}, \code{negloglik_point.val}, \code{model_par.mle_point}, \code{data.bin}
}
