% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{model_par.mle_point}
\alias{model_par.mle_point}
\title{Model Parameter Maximum Likelihood Estimation (Point Data)}
\usage{
model_par.mle_point(data, theta.init = list(a_fb = -1, tau = 1, b = 1),
  window = "full sequence")
}
\arguments{
\item{data}{a list containing all the necessary point data:
\itemize{
\item \code{seism} an earthquake catalogue data frame of parameters \code{t} (time in days) and \code{m} (magnitude)
\item \code{inj} matching injection profile data frame of parameters \code{t} (time in days), \code{dV} (flow rate in cubic
metre/day) and \code{V} (cumulative injected volume in cubic metres)
\item \code{m0} the minimum magnitude cutoff of the earthquake catalogue
\item \code{ts} shut-in time (in days) (not required for "\code{injection}")
\item \code{Tmax} upper range of the time window (in days) (not required for "\code{injection}")
\item \code{lambda0} seismicity rate (per day) at shut-in (only required for "\code{post-injection}")
}}

\item{theta.init}{an optional list of the initial values for the parameters to be optimized over
\itemize{
\item \code{a_fb} the underground feedback activation (in /cubic metre)
\item \code{tau} the mean relaxation time (in days)
\item \code{b} the slope of the Gutenberg-Richter law
}}

\item{window}{the window to be used: "\code{injection}", "\code{post-injection}", or "\code{full sequence}"}
}
\value{
a list of the parameters' MLEs:
\itemize{
\item \code{a_fb} the underground feedback activation (in /cubic metre)
\item \code{tau} the mean relaxation time (in days)
\item \code{b} the slope of the Gutenberg-Richter law
\item \code{nLL} the negative log likelihood value for the optimized parameters
}
}
\description{
Provides the maximum likelihood estimates (MLE) of the parameters of the statistical model of Mignan et al. (2017)
using the negative log likelihood function \code{negloglik_point.val()} for one of the three possible time windows (full sequence, injection phase or
post-injection phase).
}
\details{
Optimization done using the `optim()`` function of the stats package.
}
\references{
Broccardo M., Mignan A., Wiemer S., Stojadinovic B., Giardini D. (2017), Hierarchical Bayesian
Modeling of Fluid‐Induced Seismicity. Geophysical Research Letters, 44 (22), 11,357-11,367,
\href{https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/2017GL075251}{doi: 10.1002/2017GL075251}

Mignan A., Broccardo M., Wiemer S., Giardini D. (2017), Induced seismicity closed-form
traffic light system for actuarial decision-making during deep fluid injections. Sci. Rep., 7, 13607,
\href{https://www.nature.com/articles/s41598-017-13585-9}{doi: 10.1038/s41598-017-13585-9}
}
\seealso{
\code{negloglik_point.val}, \code{negloglik_hist.val}, \code{model_par.mle_hist}
}
