% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{model_posterior.distr}
\alias{model_posterior.distr}
\title{Posterior distribution estimation}
\usage{
model_posterior.distr(prior, joint_prior, LL)
}
\arguments{
\item{prior}{the list of model parameter increments as defined in \code{model_prior.distr}:
\itemize{
\item \code{bi} the vector of \code{b} increments
\item \code{ai} the vector of \code{a_fb} increments
\item \code{taui} the vector of \code{tau} increments
}}

\item{joint_prior}{the joint prior computed from \code{model_joint_prior.distr}:
\itemize{
\item \code{joint.prior_norm} the 3-dimensional array of the normalized joint prior distribution
}}

\item{LL}{the log likelihood distribution computed from \code{loglik_point.array}}
}
\value{
a list of the posterior distributions:
\itemize{
\item \code{a.post} the vector of the \code{a_fb} posterior distribution
\item \code{b.post} the vector of the \code{b} posterior distribution
\item \code{tau.post} the vector of the \code{tau} posterior distribution
\item \code{b_a.post} the array of the (\code{b}, \code{a_fb}) joint distribution
\item \code{b_tau.post} the array of the (\code{b}, \code{tau}) joint distribution
\item \code{a_tau.post} the array of the (\code{a_fb}, \code{tau}) joint distribution
\item \code{joint.post_norm} the 3-dimensional array of the normalized joint posterior distribution
}
}
\description{
Estimates the posterior distributions of the 3 parameters of the induced seismicity model of
Mignan et al. (2017), \code{model_rate.val()}, including joint posterior distributions.
}
\details{
Read Broccardo et al. (2017) for details.
}
\references{
Broccardo M., Mignan A., Wiemer S., Stojadinovic B., Giardini D. (2017), Hierarchical Bayesian
Modeling of Fluid‐Induced Seismicity. Geophysical Research Letters, 44 (22), 11,357-11,367,
\href{https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/2017GL075251}{doi: 10.1002/2017GL075251}

Mignan A., Broccardo M., Wiemer S., Giardini D. (2017), Induced seismicity closed-form
traffic light system for actuarial decision-making during deep fluid injections. Sci. Rep., 7, 13607,
\href{https://www.nature.com/articles/s41598-017-13585-9}{doi: 10.1038/s41598-017-13585-9}
}
\seealso{
\code{model_prior.distr}, \code{model_joint_prior.distr}, \code{loglik_point.array}
}
