% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{model_prior.distr}
\alias{model_prior.distr}
\title{Prior distribution estimation}
\usage{
model_prior.distr(
  par,
  ai = seq(-5, 1, 0.01),
  bi = seq(0.5, 2, 0.01),
  taui = seq(0.1, 15, 0.05)
)
}
\arguments{
\item{par}{a data frame of model parameters fitted from past stimulations
\itemize{
\item \code{b} the slope of the Gutenberg-Richter law
\item \code{a_fb} the underground feedback activation (in m^-3)
\item \code{tau} the mean relaxation time (in days)
}}
}
\value{
the list of the prior distributions for each of the 3 model parameters:
\itemize{
\item \code{bi} the vector of \code{b} increments
\item \code{b.prior} the prior Beta distribution of \code{b}
\item \code{ai} the vector of \code{a_fb} increments
\item \code{a.prior} the prior Beta distribution of \code{a_fb}
\item \code{taui} the vector of \code{tau} increments
\item \code{tau.prior} the prior Gamma distribution of \code{tau}
}
}
\description{
Estimates the prior distributions of the 3 parameters of the induced seismicity model of
Mignan et al. (2017), \code{model_rate.val()}.
}
\details{
The priors of \code{b} and \code{a_fb} are defined as Beta distributions while the prior of \code{tau} is defined as a
Gamma distribution. Read Broccardo et al. (2017) for details. The \code{par} file of Mignan et al. (2017)
(\code{par_Mignan_etal_SciRep2017.dat}) is provided as standard input. The proposed \code{bi}, \code{ai} and \code{taui} increments
are based on the ranges given in that file.
}
\references{
Broccardo M., Mignan A., Wiemer S., Stojadinovic B., Giardini D. (2017), Hierarchical Bayesian
Modeling of Fluid‐Induced Seismicity. Geophysical Research Letters, 44 (22), 11,357-11,367,
\href{https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/2017GL075251}{doi: 10.1002/2017GL075251}

Mignan A., Broccardo M., Wiemer S., Giardini D. (2017), Induced seismicity closed-form
traffic light system for actuarial decision-making during deep fluid injections. Sci. Rep., 7, 13607,
\href{https://www.nature.com/articles/s41598-017-13585-9}{doi: 10.1038/s41598-017-13585-9}
}
\seealso{
\code{model_rate.val}, \code{model_joint_prior.distr}, \code{par_Mignan_etal_SciRep2017.dat}
}
