% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{negloglik_hist.val}
\alias{negloglik_hist.val}
\title{Negative Log Likelihood Function (Histogram Data)}
\usage{
negloglik_hist.val(data, par, window = "full sequence")
}
\arguments{
\item{data}{a list containing all the necessary histogram data:
\itemize{
\item \code{seism.binned} an earthquake-rate histogram data frame of parameters \code{t} (time in days) and \code{rate} (per time bin)
\item \code{inj.binned} matching binned injection profile data frame of parameters \code{t} (time in days), \code{dV} (flow rate in cubic
metre/time bin) and \code{V} (cumulative injected volume in cubic metres)
\item \code{b} the b-value of the earthquake catalogue (not required for "\code{post-injection}")
\item \code{m0} the minimum magnitude cutoff of the earthquake catalogue (not required for "\code{post-injection}")
\item \code{ts} shut-in time (in days) (not required for "\code{injection}")
\item \code{lambda0} seismicity rate (per day) at shut-in (only required for "\code{post-injection}")
}}

\item{par}{a vector of the input parameters}

\item{window}{the window to be used: "\code{injection}", "\code{post-injection}", or "\verb{full sequence}"}
}
\value{
the negative log likelihood estimate for the specified \code{par} values
}
\description{
Estimates the negative log likelihood of the Non-Homogeneous Poisson Process (NHPP) derived from the statistical
model of Mignan et al. (2017) for histogram data and a specific set of input parameters, for three possible time
windows (full sequence, injection phase or post-injection phase).
}
\details{
Note that the function is here structured for the parameters to be optimized by \verb{optim()``, which is done automatically in }model_par.mle_hist()\verb{. The }par\verb{vector depends on the selected}window\verb{, which is dealt with in }model_par.mle_hist()` (see Details section there).
}
\references{
Mignan A., Broccardo M., Wiemer S., Giardini D. (2017), Induced seismicity closed-form
traffic light system for actuarial decision-making during deep fluid injections. Sci. Rep., 7, 13607,
\href{https://www.nature.com/articles/s41598-017-13585-9}{doi: 10.1038/s41598-017-13585-9}
}
\seealso{
\code{model_par.mle_hist}, \code{model_par.mle_point}, \code{negloglik_point.val}, \code{data.bin}
}
