% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{negloglik_point.val}
\alias{negloglik_point.val}
\title{Negative Log Likelihood Function (Point Data)}
\usage{
negloglik_point.val(data, par, window = "full sequence")
}
\arguments{
\item{data}{a list containing all the necessary point data:
\itemize{
\item \code{seism} an earthquake catalogue data frame of parameters \code{t} (time in days) and \code{m} (magnitude)
\item \code{inj} matching injection profile data frame of parameters \code{t} (time in days), \code{dV} (flow rate in cubic
metre/day) and \code{V} (cumulative injected volume in cubic metres)
\item \code{m0} the minimum magnitude cutoff of the earthquake catalogue
\item \code{ts} shut-in time (in days) (not required for "\code{injection}")
\item \code{Tmax} upper range of the time window (in days) (not required for "\code{injection}")
\item \code{lambda0} seismicity rate (per day) at shut-in (only required for "\code{post-injection}")
}}

\item{par}{a vector of the input parameters}

\item{window}{the window to be used: "\code{injection}", "\code{post-injection}", or "\code{full sequence}"}
}
\value{
the negative log likelihood estimate for the specified \code{par} values
}
\description{
Estimate the negative log likelihood of the statistical model of Mignan et al. (2017) for
a specific set of input parameters, for three possible time windows (full sequence, injection phase or
post-injection phase).
}
\details{
See Eq. A2 of Broccardo et al. (2017) for \code{'full sequence'} and Eq. A3 for \code{'injection'}.
Note that the function is here structured for the parameters to be optimized by \code{optim()``, which is done automatically in }model_par.mle_point()\code{. The }par\code{vector depends on the selected}window\code{, which is dealt with in }model_par.mle_point()` (see Details section there).
}
\references{
Broccardo M., Mignan A., Wiemer S., Stojadinovic B., Giardini D. (2017), Hierarchical Bayesian
Modeling of Fluid‐Induced Seismicity. Geophysical Research Letters, 44 (22), 11,357-11,367,
\href{https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/2017GL075251}{doi: 10.1002/2017GL075251}

Mignan A., Broccardo M., Wiemer S., Giardini D. (2017), Induced seismicity closed-form
traffic light system for actuarial decision-making during deep fluid injections. Sci. Rep., 7, 13607,
\href{https://www.nature.com/articles/s41598-017-13585-9}{doi: 10.1038/s41598-017-13585-9}
}
\seealso{
\code{model_par.mle_point}, \code{model_par.mle_hist}, \code{negloglik_hist.val}
}
