% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{par_Mignan_etal_SciRep2017}
\alias{par_Mignan_etal_SciRep2017}
\title{Induced seismicity model parameter list}
\format{
A data frame with the following columns:
\describe{
\item{site}{stimulate site}
\item{mc}{completeness magnitude}
\item{b}{slope of the Gutenberg-Richter law}
\item{a_fb}{underground feedback activation (in m^-3)}
\item{tau}{mean relaxation time (in days)}
}
}
\source{
\url{https://www.nature.com/articles/s41598-017-13585-9}
}
\usage{
par_Mignan_etal_SciRep2017
}
\description{
List of parameters \code{b}, \code{a_fb}, \code{tau} and \code{mc} estimaed for different stimulation \code{sites}, as given in
Mignan et al. (2017).
}
\references{
Mignan A., Broccardo M., Wiemer S., Giardini D. (2017), Induced seismicity closed-form
traffic light system for actuarial decision-making during deep fluid injections. Sci. Rep., 7, 13607,
\href{https://www.nature.com/articles/s41598-017-13585-9}{doi: 10.1038/s41598-017-13585-9}
}
\keyword{datasets}
