% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{poi.loglik}
\alias{poi.loglik}
\title{Poisson Log-Likelihood Function}
\usage{
poi.loglik(obs = obs, pred = pred)
}
\arguments{
\item{obs}{a vector of observed rates}

\item{pred}{a vector of predicted rates}
}
\value{
The Poisson log-likelihood estimate
}
\description{
Computes the log-likelihood of the Poisson distribution
}
