% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{ratemodel.loglik}
\alias{ratemodel.loglik}
\title{Induced Seismicity Model Log-Likelihood Function}
\usage{
ratemodel.loglik(data, theta.mle, theta.GR, t.shutin, inj = NULL)
}
\arguments{
\item{data}{The observed seismicity rate data frame
\itemize{
\item \code{t} the time (in decimal days)
\item \code{rate} the rate, or number of events at time \code{t}
}}

\item{theta.mle}{The list of model parameters
\itemize{
\item \code{a_fb} the underground feedback activation (\code{NULL} for "\code{post-injection}")
\item \code{tau} the mean relaxation time (\code{NULL} for "\code{co-injection}")
}}

\item{theta.GR}{The list of Gutenberg-Richter law parameters
\itemize{
\item \code{mc} the completeness magnitude
\item \code{b} the slope of the Gutenberg-Richter law
}}

\item{t.shutin}{The shut-in time (in decimal days)}

\item{inj}{the binned injection profile data frame with parameters (not required for "\code{post-injection}")
\itemize{
\item \code{t} the occurrence time (in decimal days)
\item \code{dV} the injected volume (in cubic metres) at \code{t}
}}
}
\value{
the model's log-likelihood estimate
}
\description{
Estimates the log-likelihood of the induced seismicity statistical
model \code{ratemodel.val()} using the Poisson formulation \code{poi.loglik()}.
}
\details{
The function automatically determines the method ("\code{co-injection}",
"\code{post-injection}", or "\code{full sequence}") from \code{data} and \code{t.shutin}.
\code{t.shutin} can be set to \code{Inf} if unknown in the future.
}
