% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{ratemodel.mle}
\alias{ratemodel.mle}
\title{Induced Seismicity model MLE}
\usage{
ratemodel.mle(data, theta.GR, t.shutin, inj = NULL)
}
\arguments{
\item{data}{The observed seismicity rate data frame
\itemize{
\item \code{t} the time (in decimal days)
\item \code{rate} the rate, or number of events at time \code{t}
}}

\item{theta.GR}{The list of Gutenberg-Richter law parameters
\itemize{
\item \code{mc} the completeness magnitude
\item \code{b} the slope of the Gutenberg-Richter law
}}

\item{t.shutin}{The shut-in time (in decimal days)}

\item{inj}{the binned injection profile data frame with parameters (not required for "\code{post-injection}")
\itemize{
\item \code{t} the occurrence time (in decimal days)
\item \code{dV} the injected volume (in cubic metres) at \code{t}
}}
}
\value{
The list of estimated parameters
\itemize{
\item \code{a_fb} the underground feedback activation (\code{NULL} for "\code{post-injection}")
\item \code{tau} the mean relaxation time (\code{NULL} for "\code{co-injection}")
}
}
\description{
Maximum Likelihood Estimation (MLE) of the induced seismicity model
parameters \code{a_fb} and \code{tau}.
}
\details{
The maximization is done on the \code{ratemodel.loglik()} result.
}
