% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{ratemodel.val}
\alias{ratemodel.val}
\title{Statistical model of induced seismicity}
\usage{
ratemodel.val(method, theta, inj = NULL, shutin = NULL,
  t.postinj = NULL)
}
\arguments{
\item{method}{COMING SOON}

\item{theta}{COMING SOON}

\item{inj}{COMING SOON}

\item{shutin}{COMING SOON}

\item{t.postinj}{COMING SOON}
}
\value{
COMING SOON
}
\description{
Description - COMING SOON
}
\details{
Details - COMING SOON
}
