% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rseismTLS.R
\docType{package}
\name{rseismTLS}
\alias{rseismTLS}
\alias{rseismTLS-package}
\title{rseismTLS: Induced Seismicity Traffic Light System}
\description{
Provides the tools to build a dynamic traffic light system (TLS)
to mitigate the risk of induced seismicity during underground stimulation
by fluid injection.
}
