% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{stat.KS_uniform}
\alias{stat.KS_uniform}
\title{Kolmogorov-Smirnov Goodness of Fit Test (uniform)}
\usage{
stat.KS_uniform(t.transf)
}
\arguments{
\item{t.transf}{a vector of transformed times}
}
\value{
the results of the K-S test as a data frame:
\itemize{
\item \code{t} the sequence of N events, from 1 to N
\item \code{lim95up} the upper 95\% K-S confidence limit
\item \code{lim95down} the lower 95\% K-S confidence limit
\item \code{lim99up} the upper 99\% K-S confidence limit
\item \code{lim99down} the lower 99\% K-S confidence limit
\item \code{boolean95} logical vector, true if events are with the 95\% K-S confidence interval
\item \code{boolean99} logical vector, true if events are with the 99\% K-S confidence interval
}
}
\description{
Kolmogorov-Smirnov Goodness of Fit Test (K-S test) to quantify the level of performance of the model
\code{model_rate.val()} and to potentially make a residual analysis (e.g., Ogata, 1988). The test is made
assuming a uniform distribution (i.e. distribution of a stationary Poisson process of intensity 1) for the
time transformed in \code{t.transform()} for the induced seismicity model.
}
\details{
See examples of induced seismicity applications in Mignan et al. (2017, 2019) and Broccardo et al. (2017).
}
\references{
Broccardo M., Mignan A., Wiemer S., Stojadinovic B., Giardini D. (2017), Hierarchical Bayesian
Modeling of Fluid‐Induced Seismicity. Geophysical Research Letters, 44 (22), 11,357-11,367,
\href{https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/2017GL075251}{doi: 10.1002/2017GL075251}

Mignan A., Broccardo M., Wiemer S., Giardini D. (2017), Induced seismicity closed-form
traffic light system for actuarial decision-making during deep fluid injections. Sci. Rep., 7, 13607,
\href{https://www.nature.com/articles/s41598-017-13585-9}{doi: 10.1038/s41598-017-13585-9}

Mignan A., Broccardo M., Wiemer S., Giardini D. (2019), Autonomous Decision-Making Against Induced
Seismicity in Deep Fluid Injections. In: Ferrari A., Laloui L. (eds), Energy Geotechnics, SEG 2018, Springer
Series in Geomechanics and Geoengineering, 369-376,
\href{https://link.springer.com/chapter/10.1007/978-3-319-99670-7_46}{doi: 10.1007/978-3-319-99670-7_46}

Ogata Y. (1988), Statistical Models for Earthquake Occurrences and Residual Analysis for Point
Processes. J. Am. Stat. Assoc.,
\href{https://amstat.tandfonline.com/doi/abs/10.1080/01621459.1988.10478560#.XfJUpZNKjOQ}{83 (401), 9-27}
}
\seealso{
\code{t.transform}, \code{model_rate.val}
}
