% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{data.bin}
\alias{data.bin}
\title{Data Binning}
\usage{
data.bin(seism, inj, tint)
}
\arguments{
\item{seism}{an earthquake catalogue data frame of parameters:
\itemize{
\item \code{t} the occurrence time (in decimal days)
\item \code{m} the magnitude
}}

\item{inj}{matching injection profile data frame of parameters:
\itemize{
\item \code{t} the occurrence time (in decimal days)
\item \code{dV} the injected volume (in cubic metres)
\item \code{V} the cumulative injected volume (in cubic metres)
}}

\item{tint}{time intervals of constant length tbin (in decimal days)}
}
\value{
A list of 2 data frames binned in time:
\itemize{
\item \code{seism.binned} with parameters \code{t} and \code{rate} (number of events per \code{tbin})
\item \code{inj.binned} with parameters \code{t}, \code{dV} (in cubic metres per \code{tbin}) and \code{V}
}
}
\description{
Bucketize the two main data sets (injection \code{inj} and seismicity \code{seism}) in time intervals \code{tint}.
}
\details{
Time \code{t} represents the centre of each bin of the time intervals \code{tint}.
}
