% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rseismTLS.R
\docType{package}
\name{rseismTLS}
\alias{rseismTLS}
\title{rseismTLS: Induced Seismicity Traffic Light System}
\description{
Provides the tools to build a dynamic traffic light system (TLS)
to mitigate the risk of induced seismicity during underground stimulation
by fluid injection.
}
\section{rseismTLS forecast functions}{

negloglik_point.val, model_par.mle_point, a_fb.val,
data.bin, model_rate.val, negloglik_hist.val, model_par.mle_hist,
t.transform, stat.KS_uniform,
model_prior.distr, model_joint_prior.distr, loglik_point.array, model_posterior.distr
}

\section{rseismNet risk functions}{
 N/A
}

